feat(activity): complete activity tracker module

+++ fitpal-springboot-app/src/main/java/com/fitpal/fitpalspringbootapp/models/User.java
@@ -51,7 +51,7 @@ public class User {
 
     private Integer dailyTargetSteps;
 
-    private Double dailyTargetActivity;
+    private Integer dailyTargetActivity;
 
     private List<FavouriteFood> favouriteFood = new ArrayList<>();
